--- libs/torus/src/lib/constants.ts
@@ -1,8 +1,0 @@
-export const torusConstants = {
-  network: 'ropsten',
-  proxyAddress: '0x4023d2a0D330bF11426B12C6144Cfb96B7fa6183',
-  discord: {
-    verifier: 'baf wallet-discord-testnet',
-    clientId: '821890148198776874',
-  }
-};
